extract metadata plugin
